Test: UserStory1to4 Integration Test
유저스토리1~4 통합 테스트
테스트파일 패키지 조정
